+++ src/lib/components/no-result/terra-no-result-notice.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { TerraButtonInterface } from '../buttons/button/data/terra-button.interface';
 
+/** @deprecated since v12. */
 @Component({
     selector: 'terra-no-result-notice',
     styleUrls: ['./terra-no-result-notice.component.scss'],