+++ CHANGELOG.md
@@ -4,6 +4,19 @@
 
 -   **terra-info** has been removed.
 
+# 12.2.0 (X.X.2021)
+
+### Minor changes
+
+-   **terra-breadcrumbs** has been deprecated.
+-   **terra-no-result-notice** has been deprecated.
+-   **terra-image-preview** use components from `@angular/material` instead of our own deprecated ones.
+-   **terra-file-input** use components from `@angular/material` instead of our own deprecated ones.
+
+### Migrations
+
+-   **migration-portlet** was added to migrate the deprecated terra-portlet to use material expansion panels instead
+
 # 12.1.1 (12.10.2021)
 
 ### Bug fixes